web-app/requirements-dev.txt
-r requirements.txt
pylint==3.2.3
black==24.4.2
pytest==8.2.2
pytest-flask==1.3.0
coverage==7.5.3
